move: types/apiのデータを一階層上に持ってきた。それに伴うimport時のpathの修正。types/envをtypes/confに移動
