+++ src/types/conf/api/ApiPath.ts
