+++ src/types/models/prefecture/Prefecture.ts
